feat(frontend): add dark mode toggle in header
