+++ frontend/src/Header.jsx
@@ -1,0 +1,35 @@
+import { useEffect, useState } from 'react'
+
+export default function Header() {
+  const [theme, setTheme] = useState('light')
+
+  useEffect(() => {
+    const stored = localStorage.getItem('theme')
+    if (stored) {
+      setTheme(stored)
+      document.documentElement.classList.toggle('dark', stored === 'dark')
+    } else if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
+      setTheme('dark')
+      document.documentElement.classList.add('dark')
+    }
+  }, [])
+
+  useEffect(() => {
+    localStorage.setItem('theme', theme)
+    document.documentElement.classList.toggle('dark', theme === 'dark')
+  }, [theme])
+
+  const toggleTheme = () => setTheme(theme === 'dark' ? 'light' : 'dark')
+
+  return (
+    <div className="flex justify-between items-center mb-6">
+      <h1 className="text-2xl font-bold">Dashboard</h1>
+      <button
+        onClick={toggleTheme}
+        className="px-3 py-1 rounded border border-gray-300 dark:border-gray-700"
+      >
+        Toggle {theme === 'dark' ? 'Light' : 'Dark'} Mode
+      </button>
+    </div>
+  )
+}
